LayoutTests/pointerevents/ios/pointer-type-attribute-on-touch-tap.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../resources/js-test.js"></script>
<script src="../../resources/ui-helper.js"></script>
<style>
body, html {
    font-family: system-ui;
    font-size: 16px;
}

.container {
    width: 200px;
    height: 200px;
    text-align: center;
    line-height: 200px;
    background: teal;
}
</style>
<script>
jsTestIsAsync = true;

receivedDown = false;
receivedUp = false;
receivedClick = false;

var downType;
var upType;
var clickType;

addEventListener("load", async () => {
    description("Verifies the expected PointerEvent.pointerType attribute when touch tapping on an element. This test requires WebKitTestRunner");

    const div = document.querySelector("#div");

    div.addEventListener("pointerdown", (event) => {
        event.preventDefault();
        receivedDown = true;
        downType = event.pointerType;
    });

    div.addEventListener("pointerup", (event) => {
        event.preventDefault();
        receivedUp = true;
        upType = event.pointerType;
    });

    div.addEventListener("click", (event) => {
        event.preventDefault();
        receivedClick = true;
        clickType = event.pointerType;
    });

    await UIHelper.tapElement(div);

    await shouldBecomeEqual("receivedDown", "true");
    await shouldBecomeEqual("receivedUp", "true");
    await shouldBecomeEqual("receivedClick", "true");

    debug("Received 'pointerdown' event with pointerType: '" + downType + "'");
    debug("Received 'pointerup' event with pointerType: '" + upType + "'");
    debug("Received 'click' event with pointerType: '" + clickType + "'");

    finishJSTest();
});
</script>
</head>
<body>
    <div class="container" id="div">Tap here</div>
</body>
</html>
